Split content-to-slot into two pieces. One that handles declarations of slots, another that handles usages.
