+++ src/collections.ts
@@ -21,7 +21,7 @@ registry.register(
     new RuleCollection('polymer-2', `Rules for projects that use Polymer 2.x`, [
       'behaviors-spelling',
       'call-super-in-callbacks',
-      'content-to-slot',
+      'content-to-slot-usages',
       'databind-with-unknown-property',
       'databinding-calls-must-be-functions',
       'dom-module-invalid-attrs',
@@ -40,7 +40,7 @@ registry.register(new RuleCollection(
 Will warn about use of deprecated Polymer 1.x features or brand new features in Polymer 2.x`,
     [
       'behaviors-spelling',
-      'content-to-slot',
+      'content-to-slot-usages',
       'databinding-calls-must-be-functions',
       'databind-with-unknown-property',
       'dom-module-invalid-attrs',
