+++ src/rules.ts
@@ -20,7 +20,8 @@ import './html/undefined-elements';
 
 import './polymer/behaviors-spelling';
 import './polymer/call-super-in-callbacks';
-import './polymer/content-to-slot';
+import './polymer/content-to-slot-declarations';
+import './polymer/content-to-slot-usages';
 import './polymer/databinding-calls-must-be-functions';
 import './polymer/databind-with-unknown-property';
 import './polymer/element-before-dom-module';
