+++ test/content-to-slot-declarations/after-fixes.html
@@ -1,0 +1,13 @@
+<dom-module id="foo-bar">
+  <template>
+    <slot></slot>
+    <slot name="foo" existing-nonsense-attribute="bar" old-content-selector=".foo">
+      Fallback content here
+    </slot>
+  </template>
+  <script>
+    Polymer({
+      is: 'foo-bar',
+    });
+  </script>
+</dom-module>
