+++ test/content-to-slot-usages/after-fixes.html
@@ -61,17 +61,3 @@
   </template>
 </paper-scroll-header-panel>
 
-
-<dom-module id="foo-bar">
-  <template>
-    <slot></slot>
-    <slot name="foo" existing-nonsense-attribute="bar" old-content-selector=".foo">
-      Fallback content here
-    </slot>
-  </template>
-  <script>
-    Polymer({
-      is: 'foo-bar',
-    });
-  </script>
-</dom-module>
