+++ test/content-to-slot-usages/before-fixes.html
@@ -61,17 +61,3 @@
   </template>
 </paper-scroll-header-panel>
 
-
-<dom-module id="foo-bar">
-  <template>
-    <content></content>
-    <content select=".foo" existing-nonsense-attribute="bar">
-      Fallback content here
-    </content>
-  </template>
-  <script>
-    Polymer({
-      is: 'foo-bar',
-    });
-  </script>
-</dom-module>
